# Optimize symbols - mmap all.