@bwplotka 2020-12-10
I was expecting more as well, but it's not bad.


```
benchstat -delta-test=none _dev/bench/store_symb_b.txt _dev/bench/store_symb_c.txt
name                                                                old time/op    new time/op    delta
TelemeterRealData_Series/alerts2w/01DN3SK96XDAEKRB1AN30AAW6E-12        1.88s ±12%     1.27s ± 0%  -32.63%
TelemeterRealData_Series/alerts15s/01DN3SK96XDAEKRB1AN30AAW6E-12       284ms ±11%     288ms ± 0%   +1.28%
TelemeterRealData_Series/subssyncs2w/01DN3SK96XDAEKRB1AN30AAW6E-12     179ms ±13%     113ms ± 0%  -36.73%
TelemeterRealData_Series/subs2w/01DN3SK96XDAEKRB1AN30AAW6E-12          5.07s ±11%     3.20s ± 0%  -36.85%
TelemeterRealData_Series/subs15s/01DN3SK96XDAEKRB1AN30AAW6E-12         782ms ±13%     907ms ± 0%  +15.97%

name                                                                old alloc/op   new alloc/op   delta
TelemeterRealData_Series/alerts2w/01DN3SK96XDAEKRB1AN30AAW6E-12        396MB ± 0%     329MB ± 0%  -16.78%
TelemeterRealData_Series/alerts15s/01DN3SK96XDAEKRB1AN30AAW6E-12      94.8MB ± 0%    94.2MB ± 0%   -0.66%
TelemeterRealData_Series/subssyncs2w/01DN3SK96XDAEKRB1AN30AAW6E-12    49.9MB ± 0%    39.3MB ± 0%  -21.35%
TelemeterRealData_Series/subs2w/01DN3SK96XDAEKRB1AN30AAW6E-12         1.08GB ± 0%    0.88GB ± 0%  -19.24%
TelemeterRealData_Series/subs15s/01DN3SK96XDAEKRB1AN30AAW6E-12         228MB ± 0%     228MB ± 0%   -0.00%

name                                                                old allocs/op  new allocs/op  delta
TelemeterRealData_Series/alerts2w/01DN3SK96XDAEKRB1AN30AAW6E-12        4.35M ± 0%     1.35M ± 0%  -68.87%
TelemeterRealData_Series/alerts15s/01DN3SK96XDAEKRB1AN30AAW6E-12       49.3k ± 0%     22.2k ± 0%  -55.07%
TelemeterRealData_Series/subssyncs2w/01DN3SK96XDAEKRB1AN30AAW6E-12      493k ± 0%      169k ± 0%  -65.72%
TelemeterRealData_Series/subs2w/01DN3SK96XDAEKRB1AN30AAW6E-12          11.8M ± 0%      3.2M ± 0%  -72.54%
TelemeterRealData_Series/subs15s/01DN3SK96XDAEKRB1AN30AAW6E-12         19.3k ± 0%     19.3k ± 0%   -0.12%
```

Comparing with https://github.com/thanos-io/thanos/pull/3557 now.

NOTE: This should not increase baseline memory per each block a bit. In my 8mln case it increased from ~5MB to 13MB so not majorly.