@bwplotka 2020-12-10
BTW NEVER compare my results with yours. This has to be revealed for each env/PC separately. cc @yeya24