@bwplotka 2020-12-10
Closing in favor of https://github.com/thanos-io/thanos/pull/3557